HACK: Allow using the following format to specify ax attributes along with their aliases:
http://blablalba/bla/phone|phone_home

Splits on | and the end part is the alias associated with the ax uri.
This advertises the ax attributes correctly to openid servers.

Without these hacks rack-openid, doesn't supply an alias for ax
attributes and are rendered as ext0, ext1, ext2 ...etc These are then
subject to ambiuous interpretations by openid servers.

This is a temporary hack.